+++ src/compiler/BytecodeGenerator.h
@@ -41,8 +41,8 @@ void Emit3(MethodGenerationContext& mgenc, uint8_t bytecode, size_t idx,
 
 void EmitHALT(MethodGenerationContext& mgenc);
 void EmitDUP(MethodGenerationContext& mgenc);
-void EmitPUSHLOCAL(MethodGenerationContext& mgenc, long idx, int ctx);
-void EmitPUSHARGUMENT(MethodGenerationContext& mgenc, long idx, int ctx);
+void EmitPUSHLOCAL(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
+void EmitPUSHARGUMENT(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
 void EmitPUSHFIELD(MethodGenerationContext& mgenc, VMSymbol* field);
 void EmitPUSHBLOCK(MethodGenerationContext& mgenc, VMInvokable* block);
 void EmitPUSHCONSTANT(MethodGenerationContext& mgenc, vm_oop_t cst);
@@ -50,8 +50,8 @@ void EmitPUSHCONSTANT(MethodGenerationContext& mgenc, uint8_t literalIndex);
 void EmitPUSHCONSTANTString(MethodGenerationContext& mgenc, VMString* str);
 void EmitPUSHGLOBAL(MethodGenerationContext& mgenc, VMSymbol* global);
 void EmitPOP(MethodGenerationContext& mgenc);
-void EmitPOPLOCAL(MethodGenerationContext& mgenc, long idx, int ctx);
-void EmitPOPARGUMENT(MethodGenerationContext& mgenc, long idx, int ctx);
+void EmitPOPLOCAL(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
+void EmitPOPARGUMENT(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
 void EmitPOPFIELD(MethodGenerationContext& mgenc, VMSymbol* field);
 void EmitSEND(MethodGenerationContext& mgenc, VMSymbol* msg);
 void EmitSUPERSEND(MethodGenerationContext& mgenc, VMSymbol* msg);
