+++ src/vmobjects/VMSymbol.cpp
@@ -68,7 +68,7 @@ VMClass* VMSymbol::GetClass() const {
 }
 
 void VMSymbol::WalkObjects(walk_heap_fn walk) {
-    for (long i = 0; i < nextCachePos; i++) {
+    for (size_t i = 0; i < nextCachePos; i++) {
         cachedClass_invokable[i] = static_cast<GCClass*>(
             walk(const_cast<GCClass*>(cachedClass_invokable[i])));
         cachedInvokable[i] =
